agent/requirements.txt
livekit-agents[openai,silero]>=1.0.0
livekit>=0.11.0
python-dotenv>=1.0.0
